ci(bench): refactor tx_exec_benchmark function into new module
The tx_exec_benchmark function has been moved from the `tx` module to a new `tx_exec` module. This included updating import paths in both the function's definition and where it is used, in `bench_tx_exec.rs` and `bench_tx_exec_blk.rs`. The commit improves the organization of the codebase by placing each major functionality in its own module.
