+++ crates/rooch-benchmarks/benches/bench_tx_exec.rs
@@ -3,7 +3,7 @@
 
 use criterion::{criterion_group, criterion_main};
 use rooch_benchmarks::helper::profiled;
-use rooch_benchmarks::tx::tx_exec_benchmark;
+use rooch_benchmarks::tx_exec::tx_exec_benchmark;
 
 criterion_group! {
     name = tx_exec_bench;
