+++ crates/rooch-benchmarks/src/lib.rs
@@ -3,3 +3,4 @@
 
 pub mod helper;
 pub mod tx;
+pub mod tx_exec;
